talk-to-docs/docprocess/requirements.txt
langchain==0.0.353
google-cloud-aiplatform==1.38.1
google-cloud-logging==3.9.0
google-cloud-discoveryengine
flask[async]
pandas
gcsfs
cloud-sql-python-connector
asyncio==3.4.3 
asyncpg==0.27.0 
pgvector==0.1.8
psycopg2-binary 
absl-py==1.4.0
bs4
lxml
trafilatura
unstructured
